stream195x: move optee-os reserved memory at the beginning of DRAM
This makes it much easier to support multiple DRAM capacity variants
with the exact same binary.

optee-os needs to be loaded at a hardcoded location in RAM and by
default it's at the end of the DRAM. This does not work well with having
only one binary for two different capacities of DRAM. Thus, optee-os is
put at the beginning of the DRAM.

This means that U-Boot and all its hardcoded addresses need to be offset
by the size of optee-os (32MB or 0x02000000) with the exception of
CONFIG_SYS_SDRAM_BASE which absolutely needs to be the actual start of
DRAM.

U-Boot has many assumptions on having the beginning of the DRAM free for
its allocations and that optee-os reserved area is at the end. The magic
used to pass the DRAM memory region to the kernel is based on such
assumptions, therefore the code resulting from such assumptions is
"fixed"/overridden by ft_board_setup() right before booting into the
kernel.

Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>