board: streamunlimited: stream195x: runtime detection of 8Gb/4Gb DRAM
This makes it possible to use the exact same binary for 8Gb or 4Gb of
DRAM variants of stream195x.

The only downside is that the U-Boot binary will only report (and use)
4Gb of DRAM, even when 8Gb is stuffed.

It'll however report the correct size to the kernel which will be able
to use the whole DRAM capacity.

Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>