include: configs: sue_fwupdate_common: quote variables in test conditions
When variables aren't defined, for example after running
`env default -a`, the `test` condition always succeeds.

When number conditions (-eq, -lt, -ge, ...) are used with a quoted
variable, if the variable isn't defined, it apparently is handled like a
zero.

When string conditions (=, <, >=, ...) are used with a quoted variable,
if the variable isn't defined, it apparently is handled like the empty
string (and will therefore not match the other operand).

Quoting variables is only required for those variables that aren't set
during the execution of the same commands in which they are used.

Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>