board: sue: common: carrier_generic_board: lower min waiting time for entering recovery mode
On S1955eMMC, the time to enter the active loop on the NPB_IN line (used
to determine if the system needs to enter recovery mode) is slightly
longer than on S1955nand and never fulfills the minimum of 3s.

It was agreed to update the Product Specification in the following way:
1) for normal boot the line is held for 1.5 to 2 seconds.
=> in reality line can be released even earlier, if host detects sign of
   life from the module (HOST_N_IRQ pin)
2) for boot into the recover mode the line is held for 5-10 seconds

As the MCU is usually holding the line for exactly 5s when it wants us to
enter recovery mode and U-Boot cannot know how long the MCU has held the
line before checking the NPB_IN, we take advantage of the gray zone and
decrease the minimum wait time from 3s to 2.2s. This makes sure we never
by accident enter recovery mode, but still have exactly 2.8 seconds of room
for us to enter active waiting loop.

Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>
[improved commit message]
Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>