board: sue: device_interface: add S1832 L4 module versions [STREAMEIGHT-4948][STREAMEIGHT-5377]
All these versions use new PMIC AXP313A.

It was found after sample production, that J108 must be removed for all
versions as it is on the bottom side of the module and outside of permitted
area where components can be placed from the bottom of the module (carrier
board must provide a hole of at least this size).

Therefore ID0 is unknown!
module_code of the S1832A 4.4W is 0x04 or 0x05
module_code of the S1832A 2.2W is 0x06 or 0x07

Signed-off-by: Wei Chen <wei.chen@streamunlimited.com>
[fix whitespace before comment]
[squashed with Wei's "add S1832A 4.4W MP module variant"]
[squashed with Wei's "add S1832A 2.2W module version"]
[written detailed commit message]
Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>