board: sue: common: enable button_sense internal pull up for Stream810X3 L5
We seem to have removed the pull-up resistor on GPIO3_21, which is
unintendedly used as "BUTTON_SENSE".

Background is that u-boot is sensing this pin to determine if recovery
image boot is requested by host. If pull-up is only enabled when kernel
boots, it will never load recovery image unless
 a) module is in factory default state
 b) SW update started via web page or
 c) uboot environment has been cleared

Signed-off-by: Wei Chen <wei.chen@streamunlimited.com>