Revert "board: sue: stream1832: set SYS_PLL back to 1344 MHz after the PMIC init [STREAMEIGHT-3745]" [STREAMEIGHT-5428]
This reverts commit 3f801050848c5f2e2bb1ad703aca93f784f262aa.

The reverted commit is one of the three commits solving STREAMEIGHT-3745.
It was identified, that this change of frequency rarely stops the boot.
This is the least important of the three commits - it restores the
frequency to 1344 MHz already in u-boot. Eventually kernel governor kicks
in and sets frequency we want. Only advantage of having faster frequency
set already in u-boot is that boot time is one second shorter.

Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>