board: sue: s195x: implement ATF / OP-TEE anti-rollback protection [STREAMEIGHT-5267]
NXP uses FIT image not only for holding kernel, but also for holding ATF,
OP-TEE, u-boot. This implements anti-rollback protection for the "first"
FIT image - the one, which is loaded by SPL.

It was decided to only implement anti-rollback protection for components,
which are running in the secure world (= ATF + OP-TEE).

Signed-off-by: Maros Medvec <maros.medvec@streamunlimited.com>