board: sue: fix environment when booting over USB
Previously environment was empty when booting from USB because
aml_try_factory_usb_burning() stopped the init flow which caused
fwupdate_init() to not be called.

This is an issue because fwupdate_init() sets the module_version which
is used to  select the correct fitImage config when booting over USB
in the factory.

This resulted in always booting the else branch of the bmfg command
which as a result always boots the new PMIC configuration.

To fix this, aml_try_factory_usb_burning() is moved to the end of
board_late_init() where all env is set up correctly.

However module_version can be wrongly set still if it is loaded from the
env, like bootlimit, but this is not correct. The variable module_version
should always be set by the module detection mechanism and it should overwrite
the saved one.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[fix comment coding style]
[move "fix potential buffer overflows" to separate commit]
Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>