env: add missing CONFIG_ENV_SIZE_REDUND option
In commit 51a80a803ac7c849229fdf03032ddb0728e30a5c Realtek added support
for the environment in an MTD device. The documentation for the
CONFIG_ENV_IS_IN_MTD option clearly states that there is supposed to be
a CONFIG_ENV_SIZE_REDUND define, however it seems that this was missed
to be added by Realtek so re-add it here.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>