tools: env: handle env and const sector size dynamically [STREAMEIGHT-6140]
Amlogic kernel NAND driver extends env and const partition with extra NAND
eraseblocks if it detects factory bad blocks in these partitions,
so they always have exactly 2 good eraseblocks. After this commit we can
take advantage of this becuase when fw_*.config does not contain the number
of ENVSECTORS it is set to the available eraseblocks.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[minor update to comment]
Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>