S195xDTS: fix ENET_IRQ_N configuration [STREAMEIGHT-5948]
Due to a wrong label in the schematics GPIO1_IO08 was configured to be
ENET_IRQ_N while in reality it is a spare GPIO called ENET_SPARE.

We want to use this GPIO for boot fail detetction instead so it should only
go high when the HW watchdog is initialized. To be sure this happens, this pin
should not go high before the kernel starts booting.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>