configs: stream210: set baudrate to the default value of 115200
Currently the U-Boot does not reconfigure the serial hardware as
`plat->skip_init` is set in `drivers/rtkdrivers/serial/serial_ameba.c`
so the configuration from the previous stage will be untouched.

However it is good to have the proper default value here, in case there
will be some change to the U-Boot in the future to re-initialize the serial
hardware.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>