board: sue: common: device_interface: enable pull-down on GPIOAO_9
This is needed because module detection relies on Wi-Fi chip to be off.
If the Wi-Fi chip is on, it can supply some power to the module detection
gpios which then can falsly return 1 instead of 0 on some boards.

No need to disable pull-down because kernel will handle the rest.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>