board: sue: axp313: remove setting of DCDC1 regulator work-mode
This piece of code was just copied from the Stream1832 U-Boot
implementation, but because the `axp313_set_dcdc_workmode()` function
was buggy anyway, it was **actually** setting the DCDC3 regulator to PWM
work-mode.

The DCDC3 regulator is the `VDD_1V8` rail and we have never observed
stability issues with this rail and after aligning with the HW department,
we agreed to remove this setting anyway.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>