board: sue: device_interface: include 0x0FFF correctly in adc_map
This value was not detected properly because in get_adc_code() we use <
instead of <= for comparision. This is okay, since min value is inclusive
in the range while max is exclusive. But on the last value this is
no longer true, we need to include the max here as well.

This has caused the fill_device_info() to exit early and not detect the
module version.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>