drivers: watchdog: add amebad2 watchdog driver
This is a simple driver that configures watchdog number 4, the only one
that can be used in the non-secure world on the A32 core.

The timeout value is taken from the CONFIG_WATCHDOG_TIMEOUT_MSECS option,
the register and bit definitions were taken from the Linux driver of the
current rtl8730e-linux-sdk-3.0_r6-beta release.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>