configs: sue_fwupdate_common: check for unfinished update on boot
Do one more check on BOOT_COUNT 1: if the previous update is flagged as
incomplete in the environment (update_in_progress is 1) - try to update
again.

More specifically, this sets flags fail & update. So, the recovery logic
starts over on each power-on, as long as update_in_progress is set.

Signed-off-by: Andriy Danylovskyy <andriy.danylovskyy@streamunlimited.com>