env: allow some user environment variables when locked
On unlocked modules, we import the default environment first and then
merge it with the stored one. This ensures that all variables required
for normal boot are available, and that any user overrides apply. Keep
this part as it is.

On locked modules the logic is the opposite: default values need to be
enforced. However, we also want to allow certain variables from the
user environment. Therefore import only those, and then merge with the
default on top.

Signed-off-by: Andriy Danylovskyy <andriy.danylovskyy@streamunlimited.com>